Summary: fix, revert last changeset and use a different approach Keywords: server
This changeset reverts the last one which did fix a race condition, by locking
the job queue all along whilst depositting a job.  However, since using a
black/grey/white marking scheme for jobs on the queue this approach is
effectively neglecting the grey phase (jobs obtained but not yet ready for
processing), locking the whole queue for an unnecessarily long period of time.

This changeset rolls back the repo to what it had been before and addresses the
culprit directly.  When dequeuing a job and encountering a grey job the
dequeuing is stopped immediately.

Furthermore, the actual dequeue routine, since of no interest at all to other
routines, is moved from the private header file to the main server component.

Signed-off-by: Sebastian Freundt <freundt@math.tu-berlin.de>